requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
PyJWT==2.1.0
pymodm==0.4.3
pymongo==3.11.4
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
Werkzeug==2.0.1
